requirements.txt
Flask==2.3.2
python-telegram-bot==21.3
requests
eth-utils
gunicorn